requirements.txt
matplotlib
numpy==1.26.4
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.4.2
scipy==1.14.0
seaborn==0.13.2
scikit-image
timm==1.0.8
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
efficientnet_pytorch==0.7.1

